Refactor product data out to common module
