+++ tests/specs/homePage.spec.ts
@@ -1,7 +1,8 @@
 import { test, expect } from '@playwright/test';
 import { HomePage, ProductItemElements } from '../pages/homePage';
-import { ExpectedText, Products, PromoBlockLinks, SwatchOutlineStyles, Colors } from '../data/homePage';
+import { ExpectedText, Products, PromoBlockLinks } from '../data/homePage';
 import { rgbToHex } from '../helpers/colorUtils';
+import { Colors, SwatchOutlineStyles } from '../data/products';
 
 const Timeouts = {
   ImageLink: 10000,
